Working Status check for bodies and SS
Added fuctionality for sorting SS from bodies, and also includes a
status check filter to check the health of bodies and stations.

It's currently limited to only 10 bodies total that will be checked for
now, this is only for development.  Due to high numbers of bodies in my
empire it slowed down testing to check them all during each test run.