Several updates
Updated captcha menus and started adding to requests that need a valid
captcha also added a variable to store if a captcha is valid or not some
methods will throw a server error if the captcha isn't valid.

Added assign assignment to spies that now works

Added view to spaceport which returns how many ships of each type are on
a planet.